% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_interval.R
\name{best_time}
\alias{best_time}
\title{best_interval - finds the best interval in early time to use for analysis - returns a time range}
\usage{
best_time(x, ports, plot = TRUE)
}
\arguments{
\item{x}{data read from dts (must be the output of "get_instantaneous" which uses "fit_convolve")}

\item{ports}{port data from the as-built MLS (must be inparticular format)}

\item{plot}{displays graphical information}

\item{dts_list}{}
}
\value{

}
\description{
best_interval - finds the best interval in early time to use for analysis - returns a time range
}
