% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_boundaries.R
\name{find_peaks}
\alias{find_peaks}
\title{find_peaks - finds the monitroing interval boundaries automatically}
\usage{
find_peaks(alltests)
}
\arguments{
\item{alltests}{data read from dts (must be the output of "get_instantaneous" which uses "fit_convolve")}
}
\value{

}
\description{
find_peaks - finds the monitroing interval boundaries automatically
}
