% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_slopes.R
\name{generate_slopes}
\alias{generate_slopes}
\title{generate_slope - generate instantaneous slopes using moving window method}
\usage{
generate_slopes(dts, ports, n = 1)
}
\arguments{
\item{dts}{data read from dts}

\item{n}{set the distance to calcualte slope}
}
\value{

}
\description{
generate_slope - generate instantaneous slopes using moving window method
}
