% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instantaneous.R
\name{get_instantaneous.list}
\alias{get_instantaneous.list}
\title{gets_instantaneous.list}
\usage{
get_instantaneous.list(dts_list, n_knots = NULL)
}
\arguments{
\item{dts_list}{data read from dts}

\item{n_knots}{specifies how to fit the data for fit_convolve()}
}
\value{

}
\description{
gets_instantaneous.list
}
