% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instantaneous.R
\name{get_instantaneous}
\alias{get_instantaneous}
\title{get_instantaneous - Gets instantaneous slopes for each point in time at every depth measurement. Takes a single dts argument or a list of dts arguments to bind output to multiple}
\usage{
get_instantaneous(x, power = 15, n_knots = NULL)
}
\arguments{
\item{x}{data read from dts}

\item{power}{specifies output power of the heating device used to calculate thermal conductivity}

\item{n_knots}{specifies how to fit the data for fit_convolve()}

\item{dts_list}{}
}
\value{

}
\description{
get_instantaneous - Gets instantaneous slopes for each point in time at every depth measurement. Takes a single dts argument or a list of dts arguments to bind output to multiple
}
