% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_time.R
\name{optimal_time}
\alias{optimal_time}
\title{optimal_time - finds the time to choose in early data based on the max difference between sand and bentonite slopes}
\usage{
optimal_time(x)
}
\arguments{
\item{x}{data read from dts (must be the output of "get_instantaneous" which uses "fit_convolve")}
}
\value{

}
\description{
optimal_time - finds the time to choose in early data based on the max difference between sand and bentonite slopes
}
