% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_heating.R
\name{process_heating}
\alias{process_heating}
\title{process_heating}
\usage{
process_heating(
  x,
  TOC,
  BOC = 70,
  resample = FALSE,
  ultima = NULL,
  set_back = 0,
  heating = "heating",
  buffer = 0.05
)
}
\arguments{
\item{x}{data read from dts}

\item{TOC}{the depth of top of casing to truncate the borehole}

\item{BOC}{the depth of bottom of casting to truncate the borehole}

\item{resample}{set to TRUE if we want to interpolate the distance to another dts test for comparison, ussually a higher resolution test}

\item{ultima}{specifies the vector or dts dataset to resample to}

\item{set_back}{allow you to set the start of heating before the actual start to see the initial temperature rise}
}
\value{

}
\description{
process_heating
}
