% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snapshots.R
\name{read_snapshots}
\alias{read_snapshots}
\title{read_snaphots}
\usage{
read_snapshots(
  x,
  power = 15,
  smooth = FALSE,
  out = "TC",
  heating = "heating",
  duration
)
}
\arguments{
\item{x}{data read from dts}

\item{power}{is the power output of the heating unit to calculate thermal conductivity}

\item{smooth}{determines if data is smoothed depending on noise. default is false}

\item{out}{provides 3 output options; 1. 'dtsobject', 2. 'temp' and 2. 'TC' or thermal_conductivity}

\item{duration}{}

\item{start}{determines start time to subset data to. default as 0, referring to 0 seconds elapsed time after heating}

\item{end}{determines end time to subset data to. default as 36000, referring to 36000 seconds elapsed time after heating}
}
\value{

}
\description{
read_snaphots
}
