% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_slopes.R
\name{generate_slopes}
\alias{generate_slopes}
\title{generate_slope - generate instantaneous slopes using moving window method}
\usage{
generate_slopes(dts, ports)
}
\arguments{
\item{dts}{data read from dts (must be the output of "get_instantaneous" which uses "fit_convolve")}
}
\value{

}
\description{
generate_slope - generate instantaneous slopes using moving window method
}
