% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instantaneous.R
\name{get_instantaneous.list}
\alias{get_instantaneous.list}
\title{get_instantaneous.list}
\usage{
get_instantaneous.list(dts_list, ports, n_knots = 12)
}
\arguments{
\item{dts_list}{list of dts data from different snapshots}

\item{n_knots}{specifies how to fit the data for fit_convolve()}
}
\value{

}
\description{
get_instantaneous.list
}
